// lesson-4/typechain-types/Caller.ts
export interface CallerInterface extends Interface {
  getFunction(
    nameOrSignature:
      | "callerValue"
      | "executeCall"
      | "executeDelegateCall"
      | "getCallerValue"
      | "getLibraryValue"
      | "libraryContract"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "CallExecuted" | "DelegateCallExecuted"
  ): EventFragment;

  encodeFunctionData(
    functionFragment: "callerValue",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "executeCall",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "executeDelegateCall",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "getCallerValue",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "getLibraryValue",
    values?: undefined
  ): string;
  encodeFunctionData(
    functionFragment: "libraryContract",
    values?: undefined
  ): string;

  decodeFunctionResult(
    functionFragment: "callerValue",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "executeCall",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "executeDelegateCall",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getCallerValue",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "getLibraryValue",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "libraryContract",
    data: BytesLike
  ): Result;
}

export namespace CallExecutedEvent {
  export type InputTuple = [
    libraryValue: BigNumberish,
    callerValue: BigNumberish
  ];
  export type OutputTuple = [libraryValue: bigint, callerValue: bigint];
  export interface OutputObject {
    libraryValue: bigint;
    callerValue: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace DelegateCallExecutedEvent {
  export type InputTuple = [
    libraryValue: BigNumberish,
    callerValue: BigNumberish
  ];
  export type OutputTuple = [libraryValue: bigint, callerValue: bigint];
  export interface OutputObject {
    libraryValue: bigint;
    callerValue: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface Caller extends BaseContract {
  connect(runner?: ContractRunner | null): Caller;
  waitForDeployment(): Promise<this>;

  interface: CallerInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  callerValue: TypedContractMethod<[], [bigint], "view">;

  executeCall: TypedContractMethod<
    [_value: BigNumberish],
    [void],
    "nonpayable"
  >;

  executeDelegateCall: TypedContractMethod<
    [_value: BigNumberish],
    [void],
    "nonpayable"
  >;

  getCallerValue: TypedContractMethod<[], [bigint], "view">;

  getLibraryValue: TypedContractMethod<[], [bigint], "view">;

  libraryContract: TypedContractMethod<[], [string], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "callerValue"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "executeCall"
  ): TypedContractMethod<[_value: BigNumberish], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "executeDelegateCall"
  ): TypedContractMethod<[_value: BigNumberish], [void], "nonpayable">;
  getFunction(
    nameOrSignature: "getCallerValue"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "getLibraryValue"
  ): TypedContractMethod<[], [bigint], "view">;
  getFunction(
    nameOrSignature: "libraryContract"
  ): TypedContractMethod<[], [string], "view">;

  getEvent(
    key: "CallExecuted"
  ): TypedContractEvent<
    CallExecutedEvent.InputTuple,
    CallExecutedEvent.OutputTuple,
    CallExecutedEvent.OutputObject
  >;
  getEvent(
    key: "DelegateCallExecuted"
  ): TypedContractEvent<
    DelegateCallExecutedEvent.InputTuple,
    DelegateCallExecutedEvent.OutputTuple,
    DelegateCallExecutedEvent.OutputObject
  >;

  filters: {
    "CallExecuted(uint256,uint256)": TypedContractEvent<
      CallExecutedEvent.InputTuple,
      CallExecutedEvent.OutputTuple,
      CallExecutedEvent.OutputObject
    >;
    CallExecuted: TypedContractEvent<
      CallExecutedEvent.InputTuple,
      CallExecutedEvent.OutputTuple,
      CallExecutedEvent.OutputObject
    >;

    "DelegateCallExecuted(uint256,uint256)": TypedContractEvent<
      DelegateCallExecutedEvent.InputTuple,
      DelegateCallExecutedEvent.OutputTuple,
      DelegateCallExecutedEvent.OutputObject
    >;
    DelegateCallExecuted: TypedContractEvent<
      DelegateCallExecutedEvent.InputTuple,
      DelegateCallExecutedEvent.OutputTuple,
      DelegateCallExecutedEvent.OutputObject
    >;
  };
}